+++ src/model.rs
@@ -27,6 +27,7 @@ pub struct Settings {
     pub white_plot_mode: bool,
     pub log_level: LogLev,
     pub open_plot_file: bool,
+    pub disable_instant_flushing: bool,
     pub use_web_gl: bool,
 }
 
@@ -43,6 +44,7 @@ impl From<Cli> for Settings {
             white_plot_mode: cli.white_plot_mode,
             log_level: cli.log_level,
             open_plot_file: cli.open_plot_file,
+            disable_instant_flushing: cli.disable_instant_flushing,
             use_web_gl: true, // TODO: add this to CLI - need to check if this works
         }
     }