+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Version 0.3.0 - ?
 - Create backup of data file if already exists
+- Add instant flushing of data file
 
 ## Version 0.2.0 - 09.07.2023
 - Added CLI